+++ src/dailiespost.ts
@@ -1,4 +1,4 @@
-import { PostFreeFightQuest } from "./tasks/postFreeFight";
+import { PostFreeFightQuest } from "./tasks/dailies/postFreeFight";
 import { runSafeGarboQuests } from "./tasks/engine";
 
 export function postFreeFightDailySetup(): void {
